requirements.txt
Django==5.2.4
Pillow==10.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
whitenoise==6.5.0